requirements.txt
# Core dependencies
torch
numpy
sounddevice
soundfile
ffmpeg-python

# ASR
openai-whisper

# Voice Cloning
openvoice
# (Choose one or both, depending on your approach)
# For RTVC:
webrtcvad-wheels
PyQt5
matplotlib
# For OpenVoice (if you want to try it too):
# pip install openvoice

# TTS
TTS  # Coqui TTS

# VAD (Voice Activity Detection)
silero-vad
torchaudio

# LLM
# For API-based: openai, requests
# For local: ollama (run outside Python), llama-cpp-python

# Frontend
gradio
streamlit
streamlit-webrtc
audio-recorder-streamlit

# Utilities
requests
